E. Dilekci, added a new background to the barchart menu with minor UI improvements, fixed the bug for the error popup also fixed the issue mentioned in the description, 27/03/2025, 12:02
There is a bug in the bar chart menu. When switching to the bar chart screen, the airport and state selection screen stays open. This does not affect the bar chart data, but it does affect the title. When selecting other states and airports with this bug, the title changes the names and the airports while the data shows the initial search data.